# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Pereval API",
    "description": "REST API для учёта и модерации горных перевалов",
    "version": "0.1.0"
  },
  "paths": {
    "/submitData/": {
      "post": {
        "tags": [
          "perevals"
        ],
        "summary": "Add Pereval",
        "operationId": "add_pereval_submitData__post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PerevalAddSchema"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {

                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "perevals"
        ],
        "summary": "Get Perevals By User Email",
        "operationId": "get_perevals_by_user_email_submitData__get",
        "parameters": [
          {
            "name": "user__email",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^[a-zA-Z0-9._%+-]+@[a-zA-Z0-9.-]+\\.[a-zA-Z]{2,}$",
              "title": "User  Email"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PerevalShowSchema"
                  },
                  "title": "Response Get Perevals By User Email Submitdata  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/submitData/{pereval_id}": {
      "get": {
        "tags": [
          "perevals"
        ],
        "summary": "Get Pereval By Id",
        "operationId": "get_pereval_by_id_submitData__pereval_id__get",
        "parameters": [
          {
            "name": "pereval_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Pereval Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PerevalShowSchema"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "perevals"
        ],
        "summary": "Update Pereval By Id",
        "operationId": "update_pereval_by_id_submitData__pereval_id__patch",
        "parameters": [
          {
            "name": "pereval_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Pereval Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PerevalReplaceSchema"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {

                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "CoordsSchema": {
        "properties": {
          "latitude": {
            "type": "number",
            "exclusiveMinimum": 0,
            "title": "Latitude"
          },
          "longitude": {
            "type": "number",
            "exclusiveMinimum": 0,
            "title": "Longitude"
          },
          "height": {
            "type": "integer",
            "exclusiveMinimum": 0,
            "title": "Height"
          }
        },
        "type": "object",
        "required": [
          "latitude",
          "longitude",
          "height"
        ],
        "title": "CoordsSchema"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "ImageSchema": {
        "properties": {
          "data": {
            "type": "string",
            "format": "base64",
            "title": "Data"
          },
          "title": {
            "type": "string",
            "maxLength": 150,
            "title": "Title"
          }
        },
        "type": "object",
        "required": [
          "data",
          "title"
        ],
        "title": "ImageSchema"
      },
      "LevelSchema": {
        "properties": {
          "winter": {
            "type": "string",
            "maxLength": 2,
            "title": "Winter",
            "default": ""
          },
          "summer": {
            "type": "string",
            "maxLength": 2,
            "title": "Summer",
            "default": ""
          },
          "autumn": {
            "type": "string",
            "maxLength": 2,
            "title": "Autumn",
            "default": ""
          },
          "spring": {
            "type": "string",
            "maxLength": 2,
            "title": "Spring",
            "default": ""
          }
        },
        "type": "object",
        "title": "LevelSchema"
      },
      "PerevalAddSchema": {
        "properties": {
          "beauty_title": {
            "type": "string",
            "maxLength": 250,
            "title": "Beauty Title"
          },
          "title": {
            "type": "string",
            "maxLength": 250,
            "title": "Title"
          },
          "other_titles": {
            "type": "string",
            "maxLength": 250,
            "title": "Other Titles"
          },
          "connect": {
            "type": "string",
            "title": "Connect",
            "default": ""
          },
          "add_time": {
            "type": "string",
            "title": "Add Time"
          },
          "coords": {
            "$ref": "#/components/schemas/CoordsSchema"
          },
          "level": {
            "$ref": "#/components/schemas/LevelSchema"
          },
          "images": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/ImageSchema"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Images",
            "default": []
          },
          "user": {
            "$ref": "#/components/schemas/UserSchema"
          }
        },
        "type": "object",
        "required": [
          "beauty_title",
          "title",
          "other_titles",
          "add_time"
        ],
        "title": "PerevalAddSchema"
      },
      "PerevalReplaceSchema": {
        "properties": {
          "beauty_title": {
            "type": "string",
            "maxLength": 250,
            "title": "Beauty Title"
          },
          "title": {
            "type": "string",
            "maxLength": 250,
            "title": "Title"
          },
          "other_titles": {
            "type": "string",
            "maxLength": 250,
            "title": "Other Titles"
          },
          "connect": {
            "type": "string",
            "title": "Connect",
            "default": ""
          },
          "add_time": {
            "type": "string",
            "title": "Add Time"
          },
          "coords": {
            "$ref": "#/components/schemas/CoordsSchema"
          },
          "level": {
            "$ref": "#/components/schemas/LevelSchema"
          },
          "images": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/ImageSchema"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Images",
            "default": []
          }
        },
        "type": "object",
        "required": [
          "beauty_title",
          "title",
          "other_titles",
          "add_time"
        ],
        "title": "PerevalReplaceSchema"
      },
      "PerevalShowSchema": {
        "properties": {
          "beauty_title": {
            "type": "string",
            "maxLength": 250,
            "title": "Beauty Title"
          },
          "title": {
            "type": "string",
            "maxLength": 250,
            "title": "Title"
          },
          "other_titles": {
            "type": "string",
            "maxLength": 250,
            "title": "Other Titles"
          },
          "connect": {
            "type": "string",
            "title": "Connect",
            "default": ""
          },
          "add_time": {
            "type": "string",
            "title": "Add Time"
          },
          "coords": {
            "$ref": "#/components/schemas/CoordsSchema"
          },
          "level": {
            "$ref": "#/components/schemas/LevelSchema"
          },
          "images": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/ImageSchema"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Images",
            "default": []
          },
          "user": {
            "$ref": "#/components/schemas/UserSchema"
          },
          "status": {
            "type": "string",
            "title": "Status",
            "default": "new"
          }
        },
        "type": "object",
        "required": [
          "beauty_title",
          "title",
          "other_titles",
          "add_time"
        ],
        "title": "PerevalShowSchema"
      },
      "UserSchema": {
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "title": "Email"
          },
          "fam": {
            "type": "string",
            "maxLength": 150,
            "title": "Fam"
          },
          "name": {
            "type": "string",
            "maxLength": 150,
            "title": "Name"
          },
          "otc": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 150
              },
              {
                "type": "null"
              }
            ],
            "title": "Otc",
            "default": ""
          },
          "phone": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 15
              },
              {
                "type": "null"
              }
            ],
            "title": "Phone"
          }
        },
        "type": "object",
        "required": [
          "email",
          "fam",
          "name"
        ],
        "title": "UserSchema"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}